requirements.txt
openai>=1.0,<2
python-dotenv>=1.0,<2
typer>=0.9.0
rich>=13.0.0
msal>=1.24.0
requests>=2.28.0
